+++ unlock-app/src/components/interface/Address.tsx
@@ -1,0 +1,30 @@
+import styled from 'styled-components'
+import React from 'react'
+import useEns from '../../hooks/useEns'
+
+interface Props {
+  address: string
+  className: string
+  id: string
+}
+
+const Address = ({ id, className, address }: Props) => {
+  const name = useEns({ address })
+  return (
+    <Abbrevation id={id} className={className} title={address}>
+      {name}
+    </Abbrevation>
+  )
+}
+
+Address.defaultProps = {
+  className: '',
+}
+
+const Abbrevation = styled.abbr`
+  border-bottom: none !important;
+  cursor: inherit !important;
+  text-decoration: none !important;
+`
+
+export default Address