+++ unlock-app/src/reducers/signatureReducer.js
@@ -3,7 +3,7 @@ import { SET_PROVIDER } from '../actions/provider'
 import { SET_NETWORK } from '../actions/network'
 import { SIGNED_DATA } from '../actions/signature'
 
-export const initialState = null
+export const initialState = {}
 
 const signatureReducer = (state = initialState, action) => {
   if ([SET_PROVIDER, SET_NETWORK, SET_ACCOUNT].indexOf(action.type) > -1) {
@@ -12,8 +12,11 @@ const signatureReducer = (state = initialState, action) => {
 
   if (action.type === SIGNED_DATA) {
     return {
-      data: action.data,
-      signature: action.signature,
+      ...state,
+      [action.id]: {
+        data: action.data,
+        signature: action.signature,
+      },
     }
   }