+++ unlock-js/src/__tests__/v0/purchaseKey.test.js
@@ -1,19 +1,18 @@
+import { ethers } from 'ethers'
 import * as UnlockV0 from 'unlock-abi-0'
+import abis from '../../abis'
 import utils from '../../utils'
-import Errors from '../../errors'
-import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
 import { prepWalletService, prepContract } from '../helpers/walletServiceHelper'
 
-const { FAILED_TO_PURCHASE_KEY } = Errors
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */)
+const UnlockVersion = abis.v0
 
 let walletService
 let transaction
 let transactionResult
 let setupSuccess
-let setupFail
 
 describe('v0', () => {
   describe('purchaseKey', () => {
@@ -21,6 +20,36 @@ describe('v0', () => {
     const owner = '0xab7c74abc0c4d48d1bdad5dcb26153fc8780f83e'
     const lockAddress = '0xd8c88be5e8eb88e38e6ff5ce186d764676012b0b'
     const data = 'key data'
+    const EventInfo = new ethers.utils.Interface(UnlockVersion.PublicLock.abi)
+    const encoder = ethers.utils.defaultAbiCoder
+    const receipt = {
+      logs: [
+        {
+          transactionIndex: 1,
+          blockNumber: 19759,
+          transactionHash:
+            '0xace0af5853a98aff70ca427f21ad8a1a958cc219099789a3ea6fd5fac30f150c',
+          address: lockAddress,
+          topics: [
+            EventInfo.events['Transfer(address,address,uint256)'].topic,
+            encoder.encode(
+              ['address'],
+              ['0x0000000000000000000000000000000000000000']
+            ),
+            encoder.encode(['address'], [owner]),
+            encoder.encode(['uint256'], [owner]),
+          ],
+          data: encoder.encode(
+            ['address', 'address', 'uint256'],
+            ['0x0000000000000000000000000000000000000000', owner, owner]
+          ),
+          logIndex: 0,
+          blockHash:
+            '0xcb27b74a5ff04b129b645bbcfde46fe1a221c2d341223df4ad2ca87e9864678a',
+          transactionLogIndex: 0,
+        },
+      ],
+    }
 
     async function nockBeforeEach(sendData = true) {
       nock.cleanAll()
@@ -44,12 +73,11 @@ describe('v0', () => {
       } else {
         result = callMethodData(owner, utils.utf8ToHex(''))
       }
-      const { testTransaction, testTransactionResult, success, fail } = result
+      const { testTransaction, testTransactionResult, success } = result
 
       transaction = testTransaction
       transactionResult = testTransactionResult
       setupSuccess = success
-      setupFail = fail
     }
 
     it('should invoke _handleMethodCall with the right params', async () => {
@@ -62,44 +90,21 @@ describe('v0', () => {
         Promise.resolve(transaction.hash)
       )
       const mock = walletService._handleMethodCall
-
-      await walletService.purchaseKey({
+      walletService.provider.waitForTransaction = jest.fn(() =>
+        Promise.resolve(receipt)
+      )
+      const tokenId = await walletService.purchaseKey({
         lockAddress,
         owner,
         keyPrice,
         data,
       })
-
-      expect(mock).toHaveBeenCalledWith(
-        expect.any(Promise),
-        TransactionTypes.KEY_PURCHASE
-      )
-
       // verify that the promise passed to _handleMethodCall actually resolves
       // to the result the chain returns from a sendTransaction call to createLock
       const result = await mock.mock.calls[0][0]
       await result.wait()
       expect(result).toEqual(transactionResult)
-      await nock.resolveWhenAllNocksUsed()
-    })
-
-    it('should emit an error if the transaction could not be sent', async () => {
-      expect.assertions(1)
-
-      const error = { code: 404, data: 'oops' }
-      await nockBeforeEach()
-      setupFail(error)
-
-      walletService.on('error', error => {
-        expect(error.message).toBe(FAILED_TO_PURCHASE_KEY)
-      })
-
-      await walletService.purchaseKey({
-        lockAddress,
-        owner,
-        keyPrice,
-        data,
-      })
+      expect(tokenId).toEqual(utils.bigNumberify(owner).toString())
       await nock.resolveWhenAllNocksUsed()
     })
 
@@ -108,14 +113,16 @@ describe('v0', () => {
 
       await nockBeforeEach(false)
       setupSuccess()
-
-      const result = await walletService.purchaseKey({
+      walletService.provider.waitForTransaction = jest.fn(() =>
+        Promise.resolve(receipt)
+      )
+      const tokenId = await walletService.purchaseKey({
         lockAddress,
         owner,
         keyPrice,
       })
       await nock.resolveWhenAllNocksUsed()
-      expect(result).toBe(transaction.hash)
+      expect(tokenId).toEqual(utils.bigNumberify(owner).toString())
     })
   })
 })