+++ unlock-js/src/__tests__/v02/updateKeyPrice.test.js
@@ -1,19 +1,18 @@
+import { ethers } from 'ethers'
 import * as UnlockV02 from 'unlock-abi-0-2'
+import abis from '../../abis'
 import utils from '../../utils'
-import Errors from '../../errors'
-import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
 import { prepWalletService, prepContract } from '../helpers/walletServiceHelper'
 
-const { FAILED_TO_UPDATE_KEY_PRICE } = Errors
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */)
+const UnlockVersion = abis.v02
 
 let walletService
 let transaction
 let transactionResult
 let setupSuccess
-let setupFail
 
 describe('v02', () => {
   describe('updateKeyPrice', () => {
@@ -39,13 +38,11 @@ describe('v02', () => {
         testTransaction,
         testTransactionResult,
         success,
-        fail,
       } = callMethodData(utils.toWei(keyPrice, 'ether'))
 
       transaction = testTransaction
       transactionResult = testTransactionResult
       setupSuccess = success
-      setupFail = fail
     }
 
     it('should invoke _handleMethodCall with the right params', async () => {
@@ -59,33 +56,57 @@ describe('v02', () => {
       )
       const mock = walletService._handleMethodCall
 
-      await walletService.updateKeyPrice({ lockAddress, keyPrice })
+      const EventInfo = new ethers.utils.Interface(UnlockVersion.PublicLock.abi)
+      const encoder = ethers.utils.defaultAbiCoder
+      const oldPrice = '1'
+      const newPrice = '2'
 
-      expect(mock).toHaveBeenCalledWith(
-        expect.any(Promise),
-        TransactionTypes.UPDATE_KEY_PRICE
+      walletService.provider.waitForTransaction = jest.fn(() =>
+        Promise.resolve({
+          logs: [
+            {
+              transactionIndex: 1,
+              blockNumber: 19759,
+              transactionHash:
+                '0xace0af5853a98aff70ca427f21ad8a1a958cc219099789a3ea6fd5fac30f150c',
+              address: lockAddress,
+              topics: [
+                EventInfo.events['PriceChanged(uint256,uint256)'].topic,
+                encoder.encode(
+                  ['uint256'],
+                  [utils.toRpcResultNumber(utils.toWei(oldPrice, 'ether'))]
+                ),
+                encoder.encode(
+                  ['uint256'],
+                  [utils.toRpcResultNumber(utils.toWei(newPrice, 'ether'))]
+                ),
+              ],
+              data: encoder.encode(
+                ['uint256', 'uint256'],
+                [
+                  utils.toRpcResultNumber(utils.toWei(oldPrice, 'ether')),
+                  utils.toRpcResultNumber(utils.toWei(newPrice, 'ether')),
+                ]
+              ),
+              logIndex: 0,
+              blockHash:
+                '0xcb27b74a5ff04b129b645bbcfde46fe1a221c2d341223df4ad2ca87e9864678a',
+              transactionLogIndex: 0,
+            },
+          ],
+        })
       )
 
+      const newKeyPrice = await walletService.updateKeyPrice({
+        lockAddress,
+        keyPrice,
+      })
       // verify that the promise passed to _handleMethodCall actually resolves
-      // to the result the chain returns from a sendTransaction call to createLock
+      // to the result the chain returns from a sendTransaction call to updateKeyPrice
       const result = await mock.mock.calls[0][0]
       await result.wait()
       expect(result).toEqual(transactionResult)
-      await nock.resolveWhenAllNocksUsed()
-    })
-
-    it('should emit an error if the transaction could not be sent', async () => {
-      expect.assertions(1)
-
-      const error = { code: 404, data: 'oops' }
-      await nockBeforeEach()
-      setupFail(error)
-
-      walletService.on('error', error => {
-        expect(error.message).toBe(FAILED_TO_UPDATE_KEY_PRICE)
-      })
-
-      await walletService.updateKeyPrice({ lockAddress, keyPrice })
+      expect(newKeyPrice).toEqual(newPrice)
       await nock.resolveWhenAllNocksUsed()
     })
   })