+++ unlock-js/src/__tests__/v11/withdrawFromLock.test.js
@@ -1,11 +1,9 @@
 import * as UnlockV11 from 'unlock-abi-1-1'
 import utils from '../../utils'
-import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
 import { prepWalletService, prepContract } from '../helpers/walletServiceHelper'
 
-const { FAILED_TO_WITHDRAW_FROM_LOCK } = Errors
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */)
 
@@ -13,7 +11,6 @@ let walletService
 let transaction
 let transactionResult
 let setupSuccess
-let setupFail
 
 describe('v11', () => {
   describe('withdrawFromLock', () => {
@@ -38,13 +35,11 @@ describe('v11', () => {
         testTransaction,
         testTransactionResult,
         success,
-        fail,
       } = callMethodData(utils.toWei(amount, 'ether'))
 
       transaction = testTransaction
       transactionResult = testTransactionResult
       setupSuccess = success
-      setupFail = fail
     }
 
     it('should invoke _handleMethodCall with the right params when no amount has been provided', async () => {
@@ -97,22 +92,6 @@ describe('v11', () => {
       const result = await mock.mock.calls[0][0]
       await result.wait()
       expect(result).toEqual(transactionResult)
-      await nock.resolveWhenAllNocksUsed()
-    })
-
-    it('should emit an error if the transaction cannot be sent', async () => {
-      expect.assertions(1)
-      const amount = '3'
-
-      const error = { code: 404, data: 'oops' }
-      await nockBeforeEach(amount)
-      setupFail(error)
-
-      walletService.on('error', error => {
-        expect(error.message).toBe(FAILED_TO_WITHDRAW_FROM_LOCK)
-      })
-
-      await walletService.withdrawFromLock({ lockAddress, amount })
       await nock.resolveWhenAllNocksUsed()
     })
   })